Version 3.0: C Code Integration in MATLAB and Simulink to Control an App
The goal of this demo is to showcase integration of C code in MATLAB and Simulink. A C# interface of a parking meter has some emulated HW registers that can be accessed through a shared memory area in Windows. The algorithms in MATLAB and Simulink control the C# App that is just used as a user interface without any intelligence. C code generation can be done out of the entire MATLAB and Simulink projects to build standalone executables. In the MATLAB implementation, there is also the possibility to use an App of the parking Meter done with the App Designer.

RELEASE: R2019b